+++ app/src/main/java/app/aryan447/mpvium/ui/player/controls/components/sheets/ChaptersSheet.kt
@@ -12,7 +12,6 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.hapticfeedback.HapticFeedbackType
-import androidx.compose.ui.platform.LocalHapticFeedback
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontStyle
 import androidx.compose.ui.text.font.FontWeight
@@ -22,6 +21,7 @@ import app.aryan447.mpvium.ui.theme.spacing
 import dev.vivvvek.seeker.Segment
 import `is`.xyz.mpv.Utils
 import kotlinx.collections.immutable.ImmutableList
+import app.aryan447.mpvium.ui.utils.rememberHapticFeedback
 
 @Composable
 fun ChaptersSheet(
@@ -66,7 +66,7 @@ fun ChapterTrack(
   onClick: () -> Unit,
   modifier: Modifier = Modifier,
 ) {
-  val haptic = LocalHapticFeedback.current
+  val haptic = rememberHapticFeedback()
   Row(
     modifier =
       modifier